change grade from decimal to float

+++ db/migrate/20140527151234_assessment_redesign.rb
@@ -12,7 +12,7 @@ class AssessmentRedesign < ActiveRecord::Migration
       t.text        :description
       t.integer     :position
       t.integer     :exp
-      t.decimal     :max_grade
+      t.float       :max_grade
       t.boolean     :published
 
       t.datetime    :open_at
@@ -51,7 +51,7 @@ class AssessmentRedesign < ActiveRecord::Migration
       t.references  :creator
       t.string      :title
       t.text        :description
-      t.decimal     :max_grade
+      t.float       :max_grade
       t.integer     :test_limit
 
       t.datetime  :last_commented_at
@@ -155,7 +155,7 @@ class AssessmentRedesign < ActiveRecord::Migration
       t.references  :submission
       t.references  :grader_course
       t.references  :std_course, index: true
-      t.decimal     :grade
+      t.float     :grade
       t.integer     :exp
       t.references  :exp_transaction
       t.boolean     :autograding_refresh, default: false
@@ -168,7 +168,7 @@ class AssessmentRedesign < ActiveRecord::Migration
       t.references  :answer
       t.references  :grading, index: true
       t.references  :grader_course, index: true
-      t.decimal     :grade
+      t.float     :grade
 
       t.datetime  :deleted_at
       t.timestamps
@@ -177,7 +177,7 @@ class AssessmentRedesign < ActiveRecord::Migration
     create_table :assessment_grading_logs do |t|
       t.references  :grading, index: true
       t.references  :grader_course
-      t.decimal     :grade
+      t.float     :grade
       t.integer     :exp
 
       t.datetime  :deleted_at
@@ -187,7 +187,7 @@ class AssessmentRedesign < ActiveRecord::Migration
     create_table  :assessment_answer_grading_logs do |t|
       t.references  :answer_grading, index: true
       t.references  :grader_course
-      t.decimal     :grade
+      t.float     :grade
 
       t.datetime  :deleted_at
       t.timestamps
